# backend/openapi/credence_openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Credence AI",
    "version": "1.0.0"
  },
  "paths": {
...
    "/history/{wallet}": {
      "get": {
        "tags": [
          "History"
        ],
        "summary": "Wallet History",
        "operationId": "wallet_history_history__wallet__get",
        "parameters": [
          {
            "name": "wallet",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Wallet"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/agents/ask": {
      "post": {
        "tags": [
          "AI Financial Agent Layer"
        ],
        "summary": "Ask Agent",
        "description": "Queries one of the autonomous credit advisors (Borrower, Lender, or Protocol).",
        "operationId": "ask_agent_agents_ask_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AgentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/agents/history/{wallet}": {
      "get": {
        "tags": [
          "AI Financial Agent Layer"
        ],
        "summary": "Get Conversation History",
        "description": "Retrieves previous conversation logs for a wallet.",
        "operationId": "get_conversation_history_agents_history__wallet__get",
        "parameters": [
          {
            "name": "wallet",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Wallet"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object"
                  },
                  "title": "Response Get Conversation History Agents History  Wallet  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/agents/insights/{wallet}": {
      "get": {
        "tags": [
          "AI Financial Agent Layer"
        ],
        "summary": "Get Agent Insights",
        "description": "Exposes personalized risk alerts, opportunities, and credit improvements.",
        "operationId": "get_agent_insights_agents_insights__wallet__get",
        "parameters": [
          {
            "name": "wallet",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Wallet"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AgentInsight"
                  },
                  "title": "Response Get Agent Insights Agents Insights  Wallet  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AgentInsight": {
        "properties": {
          "type": {
            "type": "string",
            "title": "Type"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "severity": {
            "type": "string",
            "title": "Severity"
          }
        },
        "type": "object",
        "required": [
          "type",
          "title",
          "description",
          "severity"
        ],
        "title": "AgentInsight"
      },
      "AgentRequest": {
        "properties": {
          "wallet": {
            "type": "string",
            "title": "Wallet"
          },
          "agent_type": {
            "type": "string",
            "title": "Agent Type"
          },
          "question": {
            "type": "string",
            "title": "Question"
          },
          "context": {
            "type": "object",
            "title": "Context",
            "default": {}
          }
        },
        "type": "object",
        "required": [
          "wallet",
          "agent_type",
          "question"
        ],
        "title": "AgentRequest"
      },
      "AgentResponse": {
        "properties": {
          "wallet": {
            "type": "string",
            "title": "Wallet"
          },
          "agent_type": {
            "type": "string",
            "title": "Agent Type"
          },
          "answer": {
            "type": "string",
            "title": "Answer"
          },
          "confidence": {
            "type": "number",
            "title": "Confidence"
          },
          "recommendations": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Recommendations"
          },
          "sources": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Sources"
          },
          "generated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Generated At"
          },
          "decision_trace": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Decision Trace"
          }
        },
        "type": "object",
        "required": [
          "wallet",
          "agent_type",
          "answer",
          "confidence",
          "recommendations",
          "sources",
          "generated_at",
          "decision_trace"
        ],
        "title": "AgentResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}